refactor(grafico): colocar interfaces junto a cada pantalla y actualizar imports
Mueve modelos y tipos desde grafico/interfaces a carpetas locales por feature
y alinea referencias en servicio, GraphQL, dashboard y lucro por funcionario.
