+++ src/app/modules/grafico/analisis-producto/interfaces/analisis-producto-modo-ranking.type.ts
