+++ src/app/modules/grafico/analisis-producto/interfaces/analisis-producto-nivel-venta.type.ts
