+++ src/app/modules/grafico/analisis-producto/interfaces/analisis-producto-producto-detalle.model.ts
