+++ src/app/modules/grafico/analisis-producto/interfaces/analisis-producto-vista.model.ts
