+++ src/app/modules/grafico/forma-pago/interfaces/forma-pago-detalle-procesado.model.ts
