+++ src/app/modules/grafico/forma-pago/interfaces/forma-pago-estadistica.model.ts
