+++ src/app/modules/grafico/forma-pago/interfaces/forma-pago-pantalla.model.ts
