+++ src/app/modules/grafico/gasto-categoria/interfaces/gasto-categoria-item.model.ts
