+++ src/app/modules/grafico/graficos/interfaces/graficos-dashboard-vista.model.ts
