+++ src/app/modules/grafico/graphql/forma-pago-estadisticas-por-sucursal.gql.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Query } from 'apollo-angular';
-import { FormaPagoEstadistica } from '../interfaces/forma-pago-estadistica.model';
+import { FormaPagoEstadistica } from '../forma-pago/interfaces/forma-pago-estadistica.model';
 import { formaPagoEstadisticasPorSucursalQuery } from './forma-pago-estadisticas-query';
 
 export interface Response {
