+++ src/app/modules/grafico/graphql/productos-mas-vendidos.gql.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Query } from 'apollo-angular';
 import gql from 'graphql-tag';
-import { ProductoVendidoEstadistica } from '../interfaces/producto-vendido-estadistica.model';
+import { ProductoVendidoEstadistica } from '../producto-vendido/interfaces/producto-vendido-estadistica.model';
 
 export interface Response {
   data: ProductoVendidoEstadistica[];
