+++ src/app/modules/grafico/producto-vendido/interfaces/producto-compra-periodo.model.ts
