+++ src/app/modules/grafico/producto-vendido/interfaces/producto-vendido-pantalla.model.ts
