+++ src/app/modules/grafico/producto-vendido/interfaces/producto-venta-periodo.model.ts
