+++ src/app/modules/grafico/venta-funcionario/interfaces/venta-funcionario-datos-grafico-procesados.model.ts
