+++ src/app/modules/grafico/venta-funcionario/interfaces/venta-funcionario-desde-lucro-tab-data.model.ts
