+++ src/app/modules/grafico/venta-mes/interfaces/ingreso-gasto-mes-acumulado.model.ts
